requirements.txt
discord.py==0.16.5
requests==2.13.0
cryptography==1.7.2
